project/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
langsmith>=0.1.21
